network_main/backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.2
bleach==6.2.0
boto3==1.40.49
botocore==1.40.49
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.9
django-appconf==1.1.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-js-asset==3.1.2
django-mptt==0.16.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
Faker==37.4.0
hiredis==3.2.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
packaging==25.0
pillow==11.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Unidecode==1.3.8
uritemplate==4.2.0
urllib3==2.6.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
django-celery-beat==2.8.1
drf-orjson-renderer==1.7.3
pytest-mock==3.15.1
gunicorn